requirements.txt
numpy==2.2.5
pillow==11.2.1
pycryptodome==3.22.0
python-dotenv==1.1.0
